Consolidate conftest.py into testing directory
